+++ strongtyping/tests/test_validator_type.py
@@ -4,15 +4,18 @@
 @created: 24.05.21
 @author: felix
 """
+import decimal
+import fractions
 import os
 import sys
 from functools import partial
-from typing import Dict, List, Tuple, Union
+from typing import Dict, Iterable, List, Tuple, Union
 
 import pytest
 
 from strongtyping.strong_typing import TypeMisMatch, match_class_typing, match_typing
-from strongtyping.strong_typing_utils import ValidationError, Validator
+from strongtyping.strong_typing_utils import ValidationError
+from strongtyping.types import IterValidator, Validator
 
 
 @pytest.mark.skipif(
@@ -226,6 +229,36 @@ def test_validator_type_with_default():
 
     with pytest.raises(ValidationError):
         foo({2: [2, 4]})
+
+    with pytest.raises(TypeError):
+        AllowedCluster = Validator[Iterable[Union[int, fractions.Fraction, decimal.Decimal]]]
+
+
+@pytest.mark.skipif(
+    bool(int(os.environ["ST_MODULES_INSTALLED"])) is True,
+    reason="module does not support Validator at the moment",
+)
+@pytest.mark.skipif(sys.version_info.minor < 9, reason="Available since 3.9")
+def test_iter_validator():
+    number = Union[int, fractions.Fraction, decimal.Decimal]
+
+    def allow_only_int_dec_frac(value: number):
+        if not value % 1 == 0:
+            raise TypeError
+        return True
+
+    AllowedCluster = IterValidator[Iterable[number], allow_only_int_dec_frac]
+
+    @match_typing
+    def cluster(val: AllowedCluster):
+        return True
+
+    with pytest.raises(TypeMisMatch):
+        cluster((1, 2, 3.5))  # non int float
+        cluster([1, 2, "abc"])  # non int str
+        cluster([1, 2, "3.5"])  # non int str
+        cluster([1, 2, decimal.Decimal("2.1")])  # not int decimal
+        cluster([1, 2, fractions.Fraction(3, 2)])  # non int fraction
 
 
 if __name__ == "__main__":